% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_product_asplit.R
\name{factor_product_asplit}
\alias{factor_product_asplit}
\title{Compute factor product of two arrays}
\usage{
factor_product_asplit(x, y)
}
\arguments{
\item{x, y}{arrays. Factors to be multiplied. One dimension for each variable
in scope. Dimensions must be named (to find common variables).}
}
\value{
an array with the factor product of x and y
}
\description{
Compute factor product of two arrays
}
