% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_product_ve.R
\name{sum_product_ve}
\alias{sum_product_ve}
\alias{sum_product_elim_var}
\title{Sum-Product Variable-Elimination algorithm}
\usage{
sum_product_ve(factors, elim = character(0))

sum_product_elim_var(factors, z)
}
\arguments{
\item{factors}{list of factors, each either a scalar or an array, see [factor_product] and examples.}

\item{elim}{(character vector) names of variables to be eliminated, in order of elimination.}

\item{z}{(character) name of variable two eliminate}
}
\value{
- `sum_product_ve` an array with the sum-product of `factors` after eliminating `elim`.

- `sum_product_ve`: a list with factors after computing the sum-product w.r.t `z`
}
\description{
Compute sum-product and marginalize out variables from factors stored as arrays.
}
\examples{

new_factor <- function(data, dim, scope) array(data, dim, setNames(vector("list", length(dim)), scope))
x <- new_factor(c(0, 1), 2, "x")
y <- new_factor(c(.25, .75, .75, .25), c(2, 2), c("y", "x"))

# joint prob
sum_product_ve(list(x, y), character(0))

# sum out "y"
sum_product_ve(list(x, y), "y")

# sum out "x"
sum_product_ve(list(x, y), "x")

# sum out both
sum_product_ve(list(x, y), c("y", "x")) # 1 (scalar)

# product of empty list is sat equal to 1
sum_product_ve(list(), "x")

# vectors without dimension attributes throws an error
\dontrun{
sum_product_ve(list(x, 2))   # x*2
sum_product_ve(list(2, 2))   # 4
sum_product_ve(list(1:2, 2)) # error
}
}
\seealso{
[factor_product(), cpquery_from_cpts()]
}
