% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_product.R
\name{factor_product}
\alias{factor_product}
\title{Title}
\usage{
factor_product(x, y)
}
\arguments{
\item{x, y}{arrays. Factors to be multiplied. One dimension for each variable
in scope. Dimensions must be named (to find common variables).}
}
\value{
an array with the factor product of x and y
}
\description{
Title
}
